Clarified description

+++ Geometry/Point location convex polygon.cpp
@@ -4,14 +4,14 @@
 {i, i} (i >= 0) - border, p == ch[i]
 {i, j} (i, j >= 0, i != j) - border, p on segment {ch[i], ch[j]}
 */
-//Requirement: ch - points of convex hull of polygon, ordered in CCW order, starting from the lowest of the leftmost points.
+//Requirement: ch - points of strict convex hull, ordered in CCW order, starting from the lowest of the leftmost points.
 //Complexity: O(log(n))
 //Correctness: Stress tested with linear version of this function.
 template<typename T>
 pair<int, int> point_location_convex_polygon(const vector<pt<T>>& ch, const pt<T>& p) {
     static constexpr pair<int, int> OUTSIDE = {-2, -2};
     static constexpr pair<int, int> INSIDE = {-1, -1};
-    const int n = ch.size();
+    const size_t n = ch.size();
     auto pis = [&](const pt<T>& p1, const pt<T>& p2, const pt<T>& p) -> bool {
         return dot(p2 - p1, p - p1) >= 0 && dot(p1 - p2, p - p2) >= 0;
     };
@@ -28,9 +28,9 @@ pair<int, int> point_location_convex_polygon(const vector<pt<T>>& ch, const pt<T
     if (p.x < ch[0].x) return OUTSIDE;
     if (cross(p - ch[0], ch[1] - ch[0]) > 0) return OUTSIDE;
     if (cross(ch.back() - ch[0], p - ch[0]) > 0) return OUTSIDE;
-    int l = 1, r = n;
+    size_t l = 1, r = n;
     while (l + 1 < r) {
-        int md = (l + r) / 2;
+        size_t md = (l + r) / 2;
         (cross(ch[md] - ch[0], p - ch[0]) >= 0 ? l : r) = md;
     }
     if (cross(ch[l] - ch[0], p - ch[0]) == 0) {
